feature/api_views (#37)

+++ api_yamdb/api/views.py
@@ -1,9 +1,10 @@
 from django.contrib.auth import get_user_model
 from django_filters.rest_framework import DjangoFilterBackend
+from django.shortcuts import get_object_or_404
 
 
 from rest_framework import status, viewsets
-from rest_framework.exceptions import NotFound, ValidationError
+from rest_framework.exceptions import ValidationError
 from rest_framework.filters import SearchFilter
 from rest_framework.pagination import PageNumberPagination
 from rest_framework.permissions import IsAuthenticated
@@ -11,6 +12,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_simplejwt.tokens import AccessToken
 
+from .constants import (NOT_FOUND_CATEGORY, NOT_FOUND_GENRE,
+                        NOT_FOUND_USER_WITH_EMAIL, NOT_FOUND_USER)
 from .permissions import (CommentReviewsPermission,
                           OnlyAdminOrSuperUserPermission,
                           SaveMethodsOrAdminPermission)
@@ -22,14 +25,14 @@ from .serializers import (CategoriesSerializer, CommentsSerializer,
                           TitlesCreateUpdateSerializer, TitlesSerializer,
                           ReviewsSerializer, SignupSerializer,
                           UserSerializer, UserMeSerializer)
-from .utils.code_utils import create_verification_code
-from .utils.fake_email_utils import send_fake_email
+from .utils.create_send_code import create_send_code
 
 User = get_user_model()
 
 
 class CategoriesViewSet(GetListCreateDelObjectMixin):
     """Вьюсет категорий"""
+
     queryset = Categories.objects.all().order_by('id')
     serializer_class = CategoriesSerializer
     filter_backends = (SearchFilter,)
@@ -41,13 +44,13 @@ class CategoriesViewSet(GetListCreateDelObjectMixin):
         if request.method == 'DELETE':
             slug_categories = self.kwargs.get('pk')
 
-            try:
-                categories = Categories.objects.get(slug=slug_categories)
-                categories.delete()
+            category = Categories.objects.filter(slug=slug_categories)
+            if category.exists():
+                category.delete()
                 return Response(status=status.HTTP_204_NO_CONTENT)
-            except Categories.DoesNotExist:
+            else:
                 return Response(
-                    data="Категория не найден",
+                    data=NOT_FOUND_CATEGORY,
                     status=status.HTTP_404_NOT_FOUND
                 )
 
@@ -57,6 +60,7 @@ class CategoriesViewSet(GetListCreateDelObjectMixin):
 
 class GenresViewSet(GetListCreateDelObjectMixin):
     """Вьюсет жанров"""
+
     queryset = Genres.objects.all().order_by('id')
     serializer_class = GenresSerializer
     filter_backends = (SearchFilter,)
@@ -68,13 +72,13 @@ class GenresViewSet(GetListCreateDelObjectMixin):
         if request.method == 'DELETE':
             slug_genres = self.kwargs.get('pk')
 
-            try:
-                genres = Genres.objects.get(slug=slug_genres)
+            genres = Genres.objects.filter(slug=slug_genres)
+            if genres.exists():
                 genres.delete()
                 return Response(status=status.HTTP_204_NO_CONTENT)
-            except Genres.DoesNotExist:
+            else:
                 return Response(
-                    data="Жанр не найден",
+                    data=NOT_FOUND_GENRE,
                     status=status.HTTP_404_NOT_FOUND
                 )
 
@@ -84,12 +88,15 @@ class GenresViewSet(GetListCreateDelObjectMixin):
 
 class TitlesViewSet(viewsets.ModelViewSet):
     """Вьюсет произведений"""
-    queryset = Title.objects.all().order_by('id')
+
+    queryset = Title.objects.all()
     filter_backends = (DjangoFilterBackend, SearchFilter)
     filterset_class = TitlesFilter
     search_fields = ('id', 'name',)
     pagination_class = PageNumberPagination
     permission_classes = [SaveMethodsOrAdminPermission, ]
+    ordering_fields = ['name', 'year', 'genre', 'category']
+    ordering = ['name']
 
     def get_serializer_class(self):
         """Возвращает необходимый сериализатор"""
@@ -102,22 +109,19 @@ class TitlesViewSet(viewsets.ModelViewSet):
 
 class ReviewsViewSet(viewsets.ModelViewSet):
     """Вьюсет отзывов"""
+
     serializer_class = ReviewsSerializer
     pagination_class = PageNumberPagination
     permission_classes = [CommentReviewsPermission, ]
 
     def get_queryset(self):
         title = self.kwargs.get('title_id')
-        queryset = Review.objects.filter(title=title).order_by('id')
-        return queryset
+        return Review.objects.filter(title=title).order_by('id')
 
     def perform_create(self, serializer):
         title = self.kwargs.get('title_id')
 
-        try:
-            title_object = Title.objects.get(id=title)
-        except Exception:
-            raise NotFound('Нет такого произведения')
+        title_object = get_object_or_404(Title, id=title)
 
         review_copy = Review.objects.filter(
             author=self.request.user, title=title_object
@@ -130,27 +134,26 @@ class ReviewsViewSet(viewsets.ModelViewSet):
 
 class CommentsViewSet(viewsets.ModelViewSet):
     """Вьюсет комментариев"""
+
     serializer_class = CommentsSerializer
     pagination_class = PageNumberPagination
     permission_classes = [CommentReviewsPermission, ]
 
     def get_queryset(self):
         reviews_id = self.kwargs.get('review_id')
-        queryset = Comments.objects.filter(
-            reviews_id=reviews_id).order_by('id')
+        queryset = Comments.objects.filter(reviews_id=reviews_id)
         return queryset
 
     def perform_create(self, serializer):
-        try:
-            review = Review.objects.get(id=self.kwargs.get('review_id'))
-        except Review.DoesNotExist:
-            raise NotFound('Отзыв не найден')
+
+        review = get_object_or_404(Review, id=self.kwargs.get('review_id'))
 
         serializer.save(author=self.request.user, reviews_id=review)
 
 
 class UserViewSet(viewsets.ModelViewSet):
     """Вьюсет пользователей"""
+
     queryset = User.objects.all().order_by('id')
     filter_backends = (DjangoFilterBackend, SearchFilter)
     search_fields = ('username',)
@@ -161,45 +164,32 @@ class UserViewSet(viewsets.ModelViewSet):
     def retrieve(self, request, *args, **kwargs):
         username = self.kwargs.get('pk')
 
-        try:
-            user = User.objects.get(username=username)
-            serializer = self.get_serializer(user)
-            return Response(serializer.data)
-        except User.DoesNotExist:
-            return Response(
-                data="Пользователь не найден",
-                status=status.HTTP_404_NOT_FOUND)
+        user = get_object_or_404(User, username=username)
+        serializer = self.get_serializer(user)
+        return Response(serializer.data)
 
     def partial_update(self, request, *args, **kwargs):
         username = self.kwargs.get('pk')
 
-        try:
-            user = User.objects.get(username=username)
-            serializer = self.get_serializer(
-                user, data=request.data, partial=True)
-            serializer.is_valid(raise_exception=True)
-            self.perform_update(serializer)
-            return Response(serializer.data)
-        except User.DoesNotExist:
-            return Response(
-                data="Пользователь не найден",
-                status=status.HTTP_404_NOT_FOUND)
+        user = get_object_or_404(User, username=username)
+        serializer = self.get_serializer(
+            user, data=request.data, partial=True
+        )
+        serializer.is_valid(raise_exception=True)
+        self.perform_update(serializer)
+        return Response(serializer.data)
 
     def destroy(self, request, *args, **kwargs):
         username = self.kwargs.get('pk')
 
-        try:
-            user = User.objects.get(username=username)
-            user.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        except User.DoesNotExist:
-            return Response(
-                data="Пользователь не найден",
-                status=status.HTTP_404_NOT_FOUND)
+        user = get_object_or_404(User, username=username)
+        user.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class UserMeAPIView(UserMeViewSetMixin):
     """Вьюсет пользователя для запроса /users/me/"""
+
     serializer_class = UserMeSerializer
     permission_classes = [IsAuthenticated, ]
 
@@ -209,6 +199,7 @@ class UserMeAPIView(UserMeViewSetMixin):
 
 class SignUpView(APIView):
     """Вью для регистарции пользователя"""
+
     def post(self, request):
         serializer = SignupSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
@@ -220,30 +211,25 @@ class SignUpView(APIView):
         user = User.objects.filter(username=username, email=email)
         if not user:
             if User.objects.filter(email=email).exists():
-                message = 'Пользователь с таким email уже существует'
+                message = NOT_FOUND_USER_WITH_EMAIL
                 return Response(message, status=status.HTTP_400_BAD_REQUEST)
             elif User.objects.filter(username=username).exists():
-                message = 'Пользователь с таким username уже существует'
+                message = NOT_FOUND_USER
                 return Response(message, status=status.HTTP_400_BAD_REQUEST)
             else:
                 user = User.objects.create(username=username,
                                            email=email)
-                confirmation_code = create_verification_code()
-                user.confirmation_code = confirmation_code
-                user.save()
-                send_fake_email(username, confirmation_code, email)
+                create_send_code(user, username, email)
                 return Response(serializer.data, status=status.HTTP_200_OK)
 
         user = User.objects.get(username=username, email=email)
-        confirmation_code = create_verification_code()
-        user.confirmation_code = confirmation_code
-        user.save()
-        send_fake_email(username, confirmation_code, email)
+        create_send_code(user, username, email)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
 
 class TokenView(APIView):
     """Вью для получениятокена пользователя"""
+
     def post(self, request):
         serializer = TokenSerializer(data=request.data)
